Organize GPT-OSS adapters into a model package
